+++ frontend/src/metabase/api/entity-id.ts
@@ -30,7 +30,7 @@ export type EntityType = typeof validEntityTypes[number];
 type TranslateEntityIdRequest = Partial<Record<EntityType, BaseEntityId[]>>;
 
 export type TranslateEntityIdResponseSuccess = {
-  status: "success";
+  status: "ok";
   id: string | number;
   type: EntityType;
 };