+++ frontend/src/metabase/querying/metrics/components/MetricEditor/types.ts
@@ -1,0 +1,1 @@
+export type MetricModalType = "create" | "leave";